simplify code
