+++ packages/web/components/Player/Section/PlayerCollab.tsx
@@ -1,32 +1,41 @@
-import { Box, Divider, HStack, SimpleGrid, Text } from '@metafam/ds';
+import { Box, SimpleGrid, Text } from '@metafam/ds';
 import { PlayerFragmentFragment } from 'graphql/autogen/types';
-import React from 'react';
-import { FaClock } from 'react-icons/fa';
+import React, { useMemo } from 'react';
+import { FaClock, FaGlobe } from 'react-icons/fa';
+import { getPlayerTimeZoneDisplay } from 'utils/dateHelpers';
 
-import { PlayerTimeZone } from '../PlayerTimeZone';
 import { PlayerHeroTile } from './PlayerHeroTile';
 
 type Props = { player: PlayerFragmentFragment };
-export const PlayerCollab: React.FC<Props> = ({ player }) => (
-  <SimpleGrid
-    columns={2}
-    gap={6}
-    divider={
-      <Divider height="3rem" color="whiteAlpha.400" orientation="vertical" />
-    }
-  >
-    <PlayerHeroTile title="Time zone">
-      <PlayerTimeZone player={player} />
-    </PlayerHeroTile>
-    <PlayerHeroTile title="Availability">
-      <HStack alignItems="baseline">
+export const PlayerCollab: React.FC<Props> = ({ player }) => {
+  const tzDisplay = useMemo(() => getPlayerTimeZoneDisplay(player.timezone), [
+    player.timezone,
+  ]);
+
+  return (
+    <SimpleGrid columns={2} gap={6}>
+      <PlayerHeroTile title="Time zone">
+        <Box width="1rem">
+          <FaGlobe color="blueLight" />
+        </Box>
+        <Text fontSize={{ base: 'md', sm: 'lg' }}>
+          {tzDisplay?.timeZone || '-'}
+        </Text>
+        {tzDisplay?.offset ? (
+          <Text fontSize={{ base: 'xs', sm: 'sm' }}>{tzDisplay?.offset}</Text>
+        ) : (
+          ''
+        )}
+        );
+      </PlayerHeroTile>
+      <PlayerHeroTile title="Availability">
         <Box width="1rem">
           <FaClock color="blueLight" width="1rem" />
         </Box>
         <Text fontSize={{ base: 'md', sm: 'lg' }} fontFamily="mono" mb="1">
           {`${player.availability_hours || '0'} h/week`}
         </Text>
-      </HStack>
-    </PlayerHeroTile>
-  </SimpleGrid>
-);
+      </PlayerHeroTile>
+    </SimpleGrid>
+  );
+};
